filter target regions to useable rgns and provide to all callers
I was filtering each caller to individual sets of callable regions,
but that ends up with them looking at different regions, and I found
that many callers have trouble with the same types of regions.
So now a standard bed file is created up front and passed down to
each caller as a usable set of regions for calling.